fix(bench): stage fixtures per call and reject zero-sized chunks
Two threads calling `ensure_generated` for the same missing fixture shared one
process-wide staging path, so both truncated and wrote it; whichever renamed
second hit ENOENT because the first had already moved the file away. Give each
call its own staging name and create it exclusively, so racing writers each
publish their own complete file — generation is deterministic, so the bytes
they publish are identical.

`--chunk-bytes 0` was accepted and handed to the writer, whose flush check then
fired after nearly every message: a 24 MiB fixture became hundreds of thousands
of chunks, far larger and slower to produce than the shape that was asked for.
Reject it in the CLI and in the writer's own validation, so library callers are
covered too.

+++ dev/mcapbench/src/main.rs
@@ -15,8 +15,10 @@ struct Cli {
     compression: CompressionKind,
     #[arg(long, default_value_t = 100, value_parser = clap::value_parser!(u8).range(0..=100))]
     select_percent: u8,
-    #[arg(long, default_value_t = 1_048_576)]
-    chunk_bytes: usize,
+    /// Zero would make the writer flush a chunk per message, so the shape is rejected
+    /// rather than silently producing hundreds of thousands of chunks.
+    #[arg(long, default_value_t = 1_048_576, value_parser = clap::value_parser!(u64).range(1..))]
+    chunk_bytes: u64,
     #[arg(long, value_enum, default_value_t=Layout::Interleaved)]
     layout: Layout,
 }
@@ -30,7 +32,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         FileShape {
             select_percent: cli.select_percent,
             compression: cli.compression,
-            chunk_bytes: cli.chunk_bytes,
+            chunk_bytes: cli.chunk_bytes as usize,
             layout: cli.layout,
         },
     )
